Develop

+++ patrimoniomanager/src/main/java/br/ufc/nuvem/patrimoniomanager/service/BemService.java
@@ -1,6 +1,7 @@
 package br.ufc.nuvem.patrimoniomanager.service;
 
 import br.ufc.nuvem.patrimoniomanager.model.Bem;
+import br.ufc.nuvem.patrimoniomanager.model.DTO.BemEditDTO;
 import br.ufc.nuvem.patrimoniomanager.repository.BemRepository;
 import br.ufc.nuvem.patrimoniomanager.repository.PatrimonioDataRepository;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +11,6 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -25,6 +25,10 @@ public class BemService {
 
     public List<Bem> findAll() {
         return bemRepository.findAll();
+    }
+
+    public boolean existsById(Long id) {
+        return bemRepository.existsById(id);
     }
 
     public List<Bem> searchBensByName(Optional<Long> userId, String name) {
@@ -56,8 +60,12 @@ public class BemService {
     public void delete(Long id) {
         Optional<Bem> bem = findBemById(id);
         if (bem.isPresent()) {
-            patrimonioDataRepository.deleteData(bem.get().getDirImagemBem());
+            if (bem.get().getDirImagemBem() != null)
+                patrimonioDataRepository.deleteData(bem.get().getDirImagemBem());
+
             bemRepository.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "not found");
         }
 
     }
@@ -67,20 +75,21 @@ public class BemService {
         return bemRepository.save(savedBem);
     }
 
-    public Bem update(Long id, Bem bem) {
-        if (bemRepository.existsById(id)) {
-            Bem foundBem = bemRepository.findById(id).get();
-            if (Objects.equals(foundBem.getUsuario().getCodigoUsuario(), bem.getUsuario().getCodigoUsuario())) {
-                return bemRepository.save(bem);
-            }
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Codigo usuario diferente");
+    public Bem update(BemEditDTO bem) {
+        Optional<Bem> bemOptional = bemRepository.findById(bem.getCodbem());
+        if (bemOptional.isPresent()) {
+            Bem foundBem = bemOptional.get();
+            foundBem.setLocalizacao(bem.getLocalizacao());
+            foundBem.setName(bem.getNome());
+            return bemRepository.save(foundBem);
         }
         throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Update without ID");
     }
 
     public Bem addFile(Long id, MultipartFile file) {
-        if (bemRepository.existsById(id)) {
-            Bem bem = bemRepository.findById(id).get();
+        Optional<Bem> bemOptional = bemRepository.findById(id);
+        if (bemOptional.isPresent()) {
+            Bem bem = bemOptional.get();
 
             bem.setDirImagemBem(patrimonioDataRepository.insertData(bem.getUsuario().getFolderName(), file));
             bem.setBemUrl(patrimonioDataRepository.getBemUrl(bem.getDirImagemBem()));
